+++ api/konnect/v1alpha1/konnect_conditions.go
@@ -210,3 +210,22 @@ const (
 	// mirroring has failed.
 	ControlPlaneMirroredReasonFailed = "Failed"
 )
+
+const (
+	// KonnectEntityAdoptedConditionType is the type of the condition that indicates
+	// the status for adopting existing entity in Konnect.
+	KonnectEntityAdoptedConditionType = "Adopted"
+	// KonnectEntityAdoptedReasonAdopted is the reason used with the Adopted condition type
+	// indicating that the adopting of the entity succeeded.
+	KonnectEntityAdoptedReasonAdopted = "Adopted"
+	// KonnectEntityAdoptedReasonFetchFailed is the reason used with the Adopted condition type
+	// indicating that the fetching of the entity failed.
+	KonnectEntityAdoptedReasonFetchFailed = "FetchFailed"
+	// KonnectEntityAdoptedReasonUIDConflict is the reason used with the Adopted condition type
+	// indicating that the entity is managed by another Kubernetes object with a different UID
+	// (checked by the label or tag of the entity).
+	KonnectEntityAdoptedReasonUIDConflict = "UIDConflict"
+	// KonnectEntityAdoptedReasonNotMatch is the reason used with the Adopted condition type
+	// indicating that the spec of the object does not match the entity in Konnect when adopted in match mode.
+	KonnectEntityAdoptedReasonNotMatch = "NotMatch"
+)